api/source/views.py
from rest_framework import viewsets

from api.source.models import Task
from api.source.serializers import TaskSerializer


class TaskViewSet(viewsets.ModelViewSet):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer




